[REFACTOR] 패키지 그룹화
-연관있는 도메인들을 그룹화
