+++ src/main/java/com/triprecord/triprecord/record/entity/RecordLike.java
@@ -1,8 +1,6 @@
-package com.triprecord.triprecord.comment;
+package com.triprecord.triprecord.record.entity;
 
 
-import com.triprecord.triprecord.global.util.EntityBaseTime;
-import com.triprecord.triprecord.record.Record;
 import com.triprecord.triprecord.user.User;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -11,6 +9,7 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
@@ -18,27 +17,26 @@ import lombok.NoArgsConstructor;
 
 @Entity
 @Getter
+@Table(name = "`like`")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Comment extends EntityBaseTime {
+public class RecordLike {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long commentId;
-
-    private String commentContent;
+    private Long likeId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "record_id")
-    private Record commentedRecord;
+    private Record likedRecord;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id")
-    private User commentedUser;
+    private User likedUser;
 
     @Builder
-    public Comment(String commentContent, Record record, User user) {
-        this.commentContent = commentContent;
-        this.commentedRecord = record;
-        this.commentedUser = user;
+    public RecordLike(Record record, User user) {
+        this.likedRecord = record;
+        this.likedUser = user;
     }
+
 }
