+++ src/main/java/com/triprecord/triprecord/schedule/entity/ScheduleDetail.java
@@ -1,7 +1,6 @@
-package com.triprecord.triprecord.recordimage;
+package com.triprecord.triprecord.schedule.entity;
 
-
-import com.triprecord.triprecord.record.Record;
+import com.triprecord.triprecord.schedule.entity.Schedule;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
@@ -9,6 +8,7 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import java.time.LocalDate;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
@@ -17,21 +17,25 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class RecordImage {
+public class ScheduleDetail {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long recordImgId;
+    private Long scheduleDetailId;
 
-    private String recordImg;
+    private LocalDate scheduleDetailDate;
+
+    private String scheduleContent;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "record_id")
-    private Record linkedRecord;
+    @JoinColumn(name = "schedule_id")
+    private Schedule linkedSchedule;
+
 
     @Builder
-    public RecordImage(String recordImg, Record record) {
-        this.recordImg = recordImg;
-        this.linkedRecord = record;
+    public ScheduleDetail(LocalDate localDate, String content, Schedule schedule) {
+        this.scheduleDetailDate = localDate;
+        this.scheduleContent = content;
+        this.linkedSchedule = schedule;
     }
 }
