+++ src/main/java/com/triprecord/triprecord/schedule/entity/ScheduleLike.java
@@ -1,6 +1,8 @@
-package com.triprecord.triprecord.scheduledetail;
+package com.triprecord.triprecord.schedule.entity;
 
-import com.triprecord.triprecord.schedule.Schedule;
+
+import com.triprecord.triprecord.schedule.entity.Schedule;
+import com.triprecord.triprecord.user.User;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
@@ -8,7 +10,6 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import java.time.LocalDate;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
@@ -17,25 +18,24 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ScheduleDetail {
+public class ScheduleLike {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long scheduleDetailId;
-
-    private LocalDate scheduleDetailDate;
-
-    private String scheduleContent;
+    private Long scheduleLikeId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "schedule_id")
-    private Schedule linkedSchedule;
+    private Schedule likedSchedule;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private User likedUser;
 
 
     @Builder
-    public ScheduleDetail(LocalDate localDate, String content, Schedule schedule) {
-        this.scheduleDetailDate = localDate;
-        this.scheduleContent = content;
-        this.linkedSchedule = schedule;
+    public ScheduleLike(Schedule schedule, User user) {
+        this.likedSchedule = schedule;
+        this.likedUser = user;
     }
 }
